toolcrawldata/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
requests>=2.31.0
openai>=1.30.0
resend>=0.7.0
lxml>=4.9.0
